JSON updates
* Migrates config to its own file
* Format now produces json based on config
* Small tweaks to documentation
* Hopefully benign update to the inspection/frame crawling logic to capture more data for json output

Format updates
--------------

Previously, this wouldn't correctly utilize the globally set
configuration:

    >>> from termlog import format, set_config
    >>> set_config(json=True)
    >>> format(f'This is a message')
    'This is a message'

We expected:

    >>> format(f'This is a message')
    {'data': 'This is a message', 'Timestamp': ...}

Crawling logic update
---------------------

Previously, this wouldn't pick up `message`:

    from termlog import echo, red

    message = 'green'
    echo(f'This is a {red(message)} message', json=True, timestamp=False)

    {'data': 'This is a green message'}

We expected:

    {'data': 'This is a green message', 'message': 'green'}

The update logic, though, required modifying a potentially sensitive
while loop to capture more ast types and their values.  I've already
noted that there may be a problem with `asdict(self)` within the
__str__ method of the Message class.  In at least one instance,
the `asdict` method crashed when one of the data elements of the
`fields` attribute could not be deepcopied, which is something
that `asdict` does internally.